Consolidate urbackup code, syntax changes
